add custom findBinary func to avoid using containers.conf file
this patch adds a custom findBinary func to stop relying on the upstream
config.FindHelperBinary which uses entries from the containers.conf
file. Macadam does not use that file.

This work is to simplify how the
search works but still maintaining compatibility with the locations used
by podman to reuse its gvproxy, vfkit, libkrun binaries.

Signed-off-by: lstocchi <lstocchi@redhat.com>